Split GC and RuntimeStorage
Moved RuntimeStorage into a separate file
Changed RuntimeStorage::open from an std::set to a sorted std::vector
(The sorting of open is handled in GC::run).
